+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Allow configurable equivalence relations when merging
 - Add option to select whether definitions should be found
 - Allow conversion of any type to a full JSON Schema
+- Record count of different alternatives in `ProductSchema`
 
 ### Changed
 - Use `oneOf` instead of `anyOf` in `ProductSchema`